requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
filelock==3.15.4
identify==2.6.0
iniconfig==2.0.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
psycopg2==2.9.9
pytest==8.3.2
pytest-django==4.8.0
PyYAML==6.0.2
sqlparse==0.5.1
virtualenv==20.26.3
